// src/api/openapi/model/deliveryNoteCreate.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DescriptiveArticleCreate } from './descriptiveArticleCreate';


export interface DeliveryNoteCreate { 
    articles: Array<DescriptiveArticleCreate>;
    assigned: boolean;
    company_address: string;
    date: string;
    delivery_address: string;
    delivery_note_reason_id: number;
    free: boolean;
    freight: boolean;
    job_id: number;
    name: string;
    number: number;
    variations: string;
    weight: string;
}