// src/api/openapi/model/deliveryNoteUpdate.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DeliveryNoteUpdate { 
    number: any | null;
    date: any | null;
    name: any | null;
    company_address: any | null;
    delivery_address: any | null;
    variations: any | null;
    weight: any | null;
    freight: any | null;
    free: any | null;
    assigned: any | null;
    articles: any | null;
    job_id: any | null;
    delivery_note_reason_id: any | null;
}