// src/api/openapi/model/descriptiveArticleCreate.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DescriptiveArticleCreate { 
    alternative: boolean;
    amount: number;
    description: string;
    descriptive_articles?: Array<DescriptiveArticleCreate>;
    discount: number;
    name: string;
    single_price: number;
    unit_id?: number | null;
    vat_id: number;
}